Handle UnsupportedOperation and read to offset in chunks
In addition to AttributeError, we also encounter cases where
UnsupportedOperation exception is raised when seek() method is called on a file
description (this is common with zipfile objects). This is now properly
handled.

We also needed to handle the case where offset was large enough to cause
trouble for Librarian, and are now reading to offset in 1024 byte chunks.