Do not use PID file as dependency
For some strange reason, when PID file is used as a dependency, make likes to
'update' it by starting another instance of FSAL. This does not seem to apply
to CoffeeScript and Compass targets, but we opted to remove this dependency in
all cases just to be sure.

Also fixed an issue where FSAL target was overwriting the PID file with a
left-over `echo $! > $pidfile` command.

Signed-off-by: Branko Vukelic <branko@outernet.is>