Do not double the bundles for 'librarian' comp
We are special-casing 'librarian' compoenent and excluding it from extension of
`assets.js_bundles` and `assets.css_bundles` keys because those are already
present in the librarian's main configuration.

While this can hardly be called a 'generic' solution, trying to find a proper
gneric solution would probably be relatively expensive, and this works well
enough.

Signed-off-by: Branko Vukelic <branko@outernet.is>