Use component name instead of component
Having to drag component objects around may increase memory usage. For example,
if we need to raise an exception related to components somewhere in the
install() call, it means that we need to relay the component object via the
registered member object. Instead, if we require just the name, then there is
no need to keep the whole object, and it can be released sooner.

Signed-off-by: Branko Vukelic <branko@outernet.is>