Implement tuner settings for L-band
This commit implements the tuner settings interface for L-band.

In L-band mode, we have a demodulator service which is expected to parse the
libarian settings file (librarian.json) and adjust the parameters before
restarting. This commit adds an L-band-specific form that takes the settings
from the user, and restarts the demodulator service.

Two new configuration keys are added:

- ondd.band: can be 'l' (lower-case L) or 'ku', determines which tuner
  interface will be used
- ondd.demod_restart_command: command that restarts the demodulator, by default
  it logs to syslog without doing anything

Most of the UI/backend elements were reused and made to change behavior/state
depending on the ondd.band configuration parameter. Some new translatable
strings were also introduced because of the new L-band-specific parameters.

Signed-off-by: Branko Vukelic <branko@outernet.is>