Allow extra kwargs for field() def
Some of the form-related defs like select, allow for useful extra keyword
arguments that are not directly passable through the field() def. This patch
modifies the field def to relay extra keyword arguments to other concrete field
defs. There is no special checking, so it is up to the devloper to ensure only
valid kwargs are used.

Signed-off-by: Branko Vukelic <branko@outernet.is>