Make superuser (possibly) double as system user
This commit introduces some code and a new setting that can cause the superuser
step to create a proper system account as well.

A new boolean configuration option, 'session.super_is_system' is introduced,
which causes the setup wizard step to take additional action when set to true.

The creation of a system user account is currently hard-wired to
/usr/sbin/adduser command (provided by Busybox on Outernet receivers), and is
not expected to work on systems that do not have a Busybox-compatible versions
of these commands. During account creation, an 'outernet' system account is
removed, which is also an assumption that holds true only on Outernet
receivers. Finally, the home directory of the new system account will always be
that of the removed 'outernet' users to simplify account creation.

Signed-off-by: Branko Vukelic <branko@outernet.is>